Move KeyMissingException -> Exceptions/KeyMissingException
