+++ src/Kitar/Dynamodb/Exceptions/KeyMissingException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Kitar\Dynamodb\Exceptions;
+
+class KeyMissingException extends \RuntimeException{}
