--- src/Kitar/Dynamodb/Model/KeyMissingException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Kitar\Dynamodb\Model;
-
-use RuntimeException;
-
-class KeyMissingException extends RuntimeException
-{
-}
